Fix code standard issues. #2

+++ Code/Thinkings/T-02-4-Inversion/Inversion.cpp
@@ -3,12 +3,13 @@
 #include "Inversion.h"
 #include <assert.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <time.h>
 #include <iostream>
 
 using namespace std;
 
-extern long long inversion_count(int* array, int start, int end);
+extern int64_t inversion_count(int* array, int start, int end);
 
 Inversion::Inversion()
 {
